core_detect::cache

Function test

pub(crate) fn test(bit: u32) -> bool

Tests the bit of the storage. If the storage has not been initialized, initializes it with the result of os::detect_features().

On its first invocation, it detects the CPU features and caches them in the CACHE global variable as an AtomicU64.

It uses the Feature variant to index into this variable as a bitset. If the bit is set, the feature is enabled, and otherwise it is disabled.

If the feature std_detect_env_override is enabled looks for the env variable RUST_STD_DETECT_UNSTABLE and uses its its content to disable Features that would had been otherwise detected.