core_detect::os

Function detect_features

pub(crate) fn detect_features() -> Initializer

Run-time feature detection on x86 works by using the CPUID instruction.

The CPUID Wikipedia page contains all the information about which flags to set to query which values, and in which registers these are reported.

The definitive references are:

Intel 64 and IA-32 Architectures Software Developer’s Manual Volume 2: Instruction Set Reference, A-Z.

AMD64 Architecture Programmer’s Manual, Volume 3: General-Purpose and System Instructions.